backend/news_collector/pyproject.toml
[project]
name = "news-collector"
version = "0.1.0"
description = "News & Text Collector Service"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "clickhouse-driver>=0.2.6",
    "boto3>=1.34.0",
    "kafka-python>=2.0.2",
    "httpx>=0.25.0",
    "apscheduler>=3.10.4",
    "feedparser>=6.0.10",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]

